+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [coming soon]
+
+- Adding 'For Product' meta details to default singular template for Storefront-based themes for WooCommerce.
+- Adding new widget: **KB: Back to Product**, making it easier to reference the product page (for the current product), from any area of the KB.
+
 ## v160801.36241
 
 - Fixed `E_NOTICE` error in Authors widget on `$product_id`.